+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [develop](https://github.com/adhearsion/adhearsion-asr)
+  * Feature: Alias `Result#response` to `#utterance`, with a deprecation warning
 
 # [v1.0.1](https://github.com/adhearsion/adhearsion-asr/compare/1.0.0...1.0.1) - [2013-09-30](https://rubygems.org/gems/adhearsion-asr/versions/1.0.1)
   * Bugfix: A menu definition's block context is now available